fix(refactor): move breaker resource counter to pkg (#11271)
