+++ pkg/breaker/resource_counter.go
@@ -1,4 +1,4 @@
-package main
+package breaker
 
 import (
 	"context"
